requirements.txt
altair==4.2.0
asttokens==2.0.8
async-generator==1.10
attrs==22.2.0
backcall==0.2.0
blinker==1.5
cachetools==5.2.0
cffi==1.15.1
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
commonmark==0.9.1
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
exceptiongroup==1.1.0
executing==1.0.0
face-detection==0.2.2
gitdb==4.0.10
GitPython==3.1.30
h11==0.14.0
idna==3.4
importlib-metadata==6.0.0
ipykernel==6.15.2
ipython==8.4.0
jedi==0.18.1
Jinja2==3.1.2
jsonschema==4.17.3
jupyter-core==4.11.1
jupyter_client==7.3.5
MarkupSafe==2.1.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.5
numpy==1.24.1
opencv-python==4.7.0.68
outcome==1.2.0
packaging==21.3
pandas==1.5.2
parso==0.8.3
pickleshare==0.7.5
Pillow==9.4.0
prompt-toolkit==3.0.31
protobuf==3.20.3
psutil==5.9.1
pure-eval==0.2.2
pyarrow==10.0.1
pycparser==2.21
pydeck==0.8.0
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
pytz==2022.7
pytz-deprecation-shim==0.1.0.post0
pywin32==304
pyzmq==23.2.1
requests==2.28.1
rich==13.0.1
selenium==4.8.0
semver==2.13.0
six==1.16.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
stack-data==0.5.0
streamlit==1.16.0
toml==0.10.2
toolz==0.12.0
torch==1.13.1
torchvision==0.14.1
tornado==6.2
traitlets==5.3.0
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.4.0
tzdata==2022.7
tzlocal==4.2
urllib3==1.26.13
validators==0.20.0
watchdog==2.2.1
wcwidth==0.2.5
wincertstore==0.2
wsproto==1.2.0
zipp==3.11.0
